selectProcess function finished with 1 limitation
limitation: the function is unable to return the correct index value if two
burstTime values are same
i.e. if burstTime[i] == burstTime[j] where i<j in this case it should
return i but it returns j. The probable solution would be to keep track
of equal burst time values and then return index.